# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Open Card Control Plane API
  version: 1.1
  description: >-
    REST JSON and persistent SSE control-plane endpoints. Product state remains owned by the
    control plane; provider and agent endpoints are versioned contracts. API clients may send
    an exact Open-Card-API-Version header or an Open-Card-API-Min-Version and
    Open-Card-API-Max-Version range. All version headers are optional for backwards
    compatibility: clients that omit them negotiate the legacy 1.0 contract. Requests with
    no common supported version receive 426 Upgrade Required.
servers:
  - url: /
paths:
...
  /api/v1/access/platform-domains:
    post:
      operationId: bindPlatformDomain
      parameters: [{$ref: '#/components/parameters/RequiredIdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [base_domain, dns_target]
              properties:
                base_domain: {type: string}
                dns_target: {type: string}
      responses:
        '201':
          description: Platform DNS and isolated wildcard certificate are verified.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DomainBindingResponse'}
        '409': {description: DNS or certificate verification failed; IP fallback remains available}
  /api/v1/access/application-domains:
    post:
      operationId: bindApplicationDomain
      parameters: [{$ref: '#/components/parameters/RequiredIdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [application_id, host, cname_target]
              properties:
                application_id: {type: string}
                host: {type: string}
                cname_target: {type: string}
      responses:
        '201':
          description: CNAME and isolated application certificate are verified.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DomainBindingResponse'}
        '409': {description: CNAME or certificate failed; no route was enabled}
  /api/v1/access/domain-routes:
    post:
      operationId: applyDomainRoutes
      parameters: [{$ref: '#/components/parameters/RequiredIdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DomainRoutesRequest'}
      responses:
        '201':
          description: Verified longest-prefix routes were atomically loaded.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AccessStateResponse'}
        '403': {description: Worker or database target is not routable}
        '409': {description: Domain/path conflict or unverified domain/certificate}
  /api/v1/access/platform-application-domains:
    post:
      operationId: createPlatformApplicationDomain
      parameters: [{$ref: '#/components/parameters/RequiredIdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [application_id, application_name, platform]
              properties:
                application_id: {type: string}
                application_name: {type: string}
                platform: {$ref: '#/components/schemas/DomainBinding'}
      responses:
        '201': {description: Stable slug plus short-ID platform hostname was created}
        '409': {description: Platform domain or wildcard certificate is not ready}
  /api/v1/access/certificates/renew:
    post:
      operationId: renewCertificate
      parameters: [{$ref: '#/components/parameters/RequiredIdempotencyKey'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [application_id, certificate]
              properties:
                application_id: {type: string}
                certificate: {$ref: '#/components/schemas/CertificateReference'}
      responses:
        '201': {description: Replacement certificate reference is ready}
        '503': {description: Renewal failed; current route and IP fallback remain active}
...
components:
  parameters:
...
    RequiredIdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
...
    RouteTarget:
      type: object
      additionalProperties: false
      required: [application_id, deployment_id, service_name, port, path, routable]
      properties:
        application_id: {type: string}
        deployment_id: {type: string}
        service_name: {type: string}
        port: {type: integer, minimum: 1, maximum: 65535}
        path: {type: string, pattern: '^/'}
        routable: {type: boolean}
...
    DomainBinding:
      type: object
      additionalProperties: false
      required: [id, kind, host, status, created_at, updated_at]
      properties:
        id: {type: string}
        kind: {type: string, enum: [platform, application]}
        application_id: {type: string}
        platform_domain_id: {type: string}
        managed: {type: boolean}
        host: {type: string}
        expected_cname: {type: string}
        status: {type: string, enum: [pending, verifying, verified, certificate_pending, ready, failed]}
        certificate_ref: {type: string}
        failure_reason: {type: string}
        verified_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CertificateReference:
      type: object
      additionalProperties: false
      required: [id, domain_binding_id, host, secret_ref, status, updated_at]
      properties:
        id: {type: string}
        domain_binding_id: {type: string}
        host: {type: string}
        secret_ref:
          type: object
          description: Opaque reference only; private material is never returned.
        status: {type: string, enum: [pending, ready, renewing, failed]}
        fingerprint: {type: string}
        not_before: {type: string, format: date-time}
        not_after: {type: string, format: date-time}
        renew_after: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    AccessState:
      type: object
      required: [runtime_ready, ip_available, https_ready, serving, serving_old_release, message]
      properties:
        application_id: {type: string}
        deployment_id: {type: string}
        runtime_ready: {type: boolean}
        ip_available: {type: boolean}
        ip_address: {type: string}
        domain_status: {type: string}
        domain: {type: string}
        https_ready: {type: boolean}
        serving: {type: boolean}
        serving_old_release: {type: boolean}
        message: {type: string}
    AccessStateResponse:
      type: object
      required: [access_state]
      properties:
        access_state: {$ref: '#/components/schemas/AccessState'}
    DomainBindingResponse:
      type: object
      required: [binding, certificate, access_state]
      properties:
        binding: {$ref: '#/components/schemas/DomainBinding'}
        certificate: {$ref: '#/components/schemas/CertificateReference'}
        access_state: {$ref: '#/components/schemas/AccessState'}
    DomainRoutesRequest:
      type: object
      additionalProperties: false
      required: [binding, certificate, targets, runtime_ready]
      properties:
        binding: {$ref: '#/components/schemas/DomainBinding'}
        certificate: {$ref: '#/components/schemas/CertificateReference'}
        targets:
          type: array
          minItems: 1
          items: {$ref: '#/components/schemas/RouteTarget'}
        runtime_ready: {type: boolean}
...